Remove the use of doubly linked list in merge two
Since using the doubly linked list will change the prev
in the node, which will be used in sorting function, it
should be avoided. Furthermore, the maintenance of doubly
linked list will use more instructions than singly linked
list. For the above reasons, it was changed to singly linked
list.

Change-Id: I377b8457ec193e1cc9cb7abffdc79207ed5a34a6