Fix incorrect use of DROP_SIZE in dudect
Previously, the code removed DROP_SIZE data from
both the head and tail in measure, but this caused
2 * DROP_SIZE measurements to be skipped. The missing
data was instead processed using uninitialized array
values, leading to unintended behavior.

This commit modifies measure to range from 1 to N_MEASURES
and drops the head and tail in differentiate, prepare_percentiles
and update_statistics.

Change-Id: I193d35f73b39a232dccfea935ae3c043e2e5255f